// src/lib/orval/generated/model/experienceTimingType.ts
/**
 * Generated by orval v7.13.1 🍺
 * Do not edit manually.
 * Gradegy API
 * Interactive documentation for the Gradegy API
 * OpenAPI spec version: 1.0
 */

/**
 * Determines how the experience availability window is calculated.
 */
export type ExperienceTimingType = (typeof ExperienceTimingType)[keyof typeof ExperienceTimingType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ExperienceTimingType = {
  delay_after_previous: 'delay_after_previous',
  start_date_and_length: 'start_date_and_length',
  date_range: 'date_range',
} as const;